rest/requirements.txt
Jinja2==3.1.6
MarkupSafe==2.1.5
PyGithub==1.59.1
python-gitlab==8.1.0
PyJWT==2.10.1
PyYAML==6.0.2
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
SQLAlchemy==1.4.54
aiohttp==3.12.15
alembic==1.16.5
apache-airflow-client==3.0.2
bcrypt==3.2.2
cryptography==45.0.7
fastapi==0.116.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
pydantic[email]==2.11.7
pytzdata==2020.1
sqlalchemy-redshift==0.8.14
tomli==2.0.1
tomli_w==1.0.0
urllib3==2.5.0
uvicorn==0.35.0
uvicorn[standard]==0.35.0